+++ packages/renderer/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useFeatureFlags';
 export * from './useIpc';
 export * from './useKeyCombo';
 export * from './useModifierKey';