+++ packages/shared/src/featureFlags.ts
@@ -1,0 +1,3 @@
+export const FeatureFlagUuids = [
+	'0196518a-ab04-74b7-b69f-98f85176382a'
+] as const;